- Saving PDB is completed in editing mode - Check all fields if they are saved in edit pending

+++ includes/templates/edit-wpunity_asset3D-saveData.php
@@ -160,7 +160,6 @@ function wpunity_create_asset_moleculeExtra_frontend($asset_newID){
 
 function wpunity_create_asset_3DFilesExtra_frontend($asset_newID, $assetTitleForm, $gameSlug){
 
-    $totalTextures = 0; //counter in order to know how much textures we have
     $textureNamesIn = [];
     $tContent = [];
 
@@ -190,13 +189,6 @@ function wpunity_create_asset_3DFilesExtra_frontend($asset_newID, $assetTitleFor
    
     $mtl_content = $_POST['mtlFileInput'];
     $obj_content = $_POST['objFileInput'];
-    
-    // Remove this debugging piece of code in the end
-//    $fh = fopen("output_mtlContent.txt", "w");
-//    fwrite($fh, $mtl_content);
-//    fclose($fh);
-    // - until here
-    
     
     // MTL : Open mtl file and replace jpg filename
     if($_POST['mtlFileInput'] && isset($_POST['textureFileInput'])) {
@@ -237,50 +229,45 @@ function wpunity_create_asset_3DFilesExtra_frontend($asset_newID, $assetTitleFor
         }
     }
     
+    
+    // PDB
+    if (isset($_POST['pdbFileInput']))
+        if (strlen($_POST['pdbFileInput'])>0) {
+            $pdbFile_id = wpunity_upload_AssetText($_POST['pdbFileInput'], 'material' . $assetTitleForm, $asset_newID, $gameSlug);
+            update_post_meta($asset_newID, 'wpunity_asset3d_pdb', $pdbFile_id);
+        }
+    
     // SCREENSHOT
-    if (isset($_POST['sshotFileInput'])) {
-        $screenShotFile_id = wpunity_upload_Assetimg64($_POST['sshotFileInput'], $assetTitleForm, $asset_newID, $gameSlug);
-        update_post_meta($asset_newID, 'wpunity_asset3d_screenimage', $screenShotFile_id);
+    if (isset($_POST['sshotFileInput']) ) {
+        if (strlen($_POST['sshotFileInput'])>0) {
+            $screenShotFile_id = wpunity_upload_Assetimg64($_POST['sshotFileInput'], $assetTitleForm, $asset_newID, $gameSlug);
+            update_post_meta($asset_newID, 'wpunity_asset3d_screenimage', $screenShotFile_id);
+        }
     }
 
 }
 
 function wpunity_copy_3Dfiles($asset_newID, $asset_sourceID){
-
-    //$fo = fopen("output_copy_asset.txt","w");
-    //fwrite($fo, print_r([$asset_newID, $assetTitleForm, $gameSlug, $asset_sourceID], true) );
-    //fclose($fo);
     
     // Get the source post
     $assetpostMeta = get_post_meta($asset_sourceID);
     
-    // and Clone all the meta fields related to 3D
-    update_post_meta($asset_newID, 'wpunity_asset3d_mtl', $assetpostMeta['wpunity_asset3d_mtl'][0]);
-    update_post_meta($asset_newID, 'wpunity_asset3d_obj', $assetpostMeta['wpunity_asset3d_obj'][0]);
-    update_post_meta($asset_newID, 'wpunity_asset3d_screenimage', $assetpostMeta['wpunity_asset3d_screenimage'][0]);
+    if ($assetpostMeta['wpunity_asset3d_pdb'][0])
+        update_post_meta($asset_newID, 'wpunity_asset3d_pdb', $assetpostMeta['wpunity_asset3d_pdb'][0]);
     
+    if ($assetpostMeta['wpunity_asset3d_mtl'][0])
+        update_post_meta($asset_newID, 'wpunity_asset3d_mtl', $assetpostMeta['wpunity_asset3d_mtl'][0]);
     
+    if($assetpostMeta['wpunity_asset3d_obj'][0])
+        update_post_meta($asset_newID, 'wpunity_asset3d_obj', $assetpostMeta['wpunity_asset3d_obj'][0]);
     
-    delete_post_meta($asset_newID, 'wpunity_asset3d_diffimage');
-    for ($m=0; $m < count($assetpostMeta['wpunity_asset3d_diffimage']); $m++)
-        add_post_meta($asset_newID, 'wpunity_asset3d_diffimage', $assetpostMeta['wpunity_asset3d_diffimage'][$m]);
+    if($assetpostMeta['wpunity_asset3d_screenimage'][0])
+        update_post_meta($asset_newID, 'wpunity_asset3d_screenimage', $assetpostMeta['wpunity_asset3d_screenimage'][0]);
     
-    
+    if (count($assetpostMeta['wpunity_asset3d_diffimage']) > 0) {
+        delete_post_meta($asset_newID, 'wpunity_asset3d_diffimage');
+        for ($m = 0; $m < count($assetpostMeta['wpunity_asset3d_diffimage']); $m++)
+            add_post_meta($asset_newID, 'wpunity_asset3d_diffimage', $assetpostMeta['wpunity_asset3d_diffimage'][$m]);
+    }
 }
-
-function wpunity_create_asset_pdbFiles_frontend($asset_newID, $assetTitleForm, $gameSlug){
-    $pdb_content = $_POST['pdbFileInput'];
-    $screenShotFile = $_POST['sshotFileInput'];
-
-    // PDB
-    $pdbFile_id = wpunity_upload_AssetText($pdb_content, 'material'.$assetTitleForm, $asset_newID, $gameSlug);
-
-    // SCREENSHOT
-    $screenShotFile_id = wpunity_upload_Assetimg64($screenShotFile, $assetTitleForm, $asset_newID, $gameSlug);
-
-    update_post_meta($asset_newID, 'wpunity_asset3d_pdb', $pdbFile_id);
-    update_post_meta($asset_newID, 'wpunity_asset3d_screenimage', $screenShotFile_id);
-}
-
-
 ?>
